apps/testing/requirements.txt
# AI Factory Agents - Dependencies

# Core
anthropic>=0.40.0
groq>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Database
supabase>=2.0.0

# HTTP Clients
httpx>=0.27.0
aiohttp>=3.9.0

# Async
asyncio>=3.4.3

# CLI
click>=8.1.0
rich>=13.0.0

# Utilities
jinja2>=3.1.0
python-dateutil>=2.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
